requirements.txt
numpy
azure-storage-blob
azure-identity
quart
quart-cors
python-dotenv
hypercorn
Pillow
cryptography
pyyaml
pydantic
python-magic
nachet-datastore @git+https://github.com/ai-cfia/nachet-datastore.git@e8b32a1f05408493e8f666e6e4872eee2cbecf45 #TODO : change the branch to main
